Move HandleSignals class
