+++ perun/utils/structs.py
@@ -7,14 +7,18 @@ from enum import Enum
 from typing import Optional, Any, TYPE_CHECKING, cast, Callable
 import enum
 import shlex
+import signal
 
 # Third-Party Imports
 
 # Perun Imports
+from perun.utils.common import common_kit
 from perun.utils.common.common_kit import ColorChoiceType, PROFILE_TRACKED, PROFILE_UNTRACKED
+from perun.utils.exceptions import SignalReceivedException
 
 if TYPE_CHECKING:
     import types
+    import traceback
 
     import numpy.typing as npt
     import numpy
@@ -592,3 +596,80 @@ class ClassificationMethod(Enum):
     FastCheck = 1
     LinearRegression = 2
     PolynomialRegression = 3
+
+
+class HandledSignals:
+    """Context manager for code blocks that need to handle one or more signals during their
+    execution.
+
+    The CM offers a default signal handler and a default handler exception. In this scenario, the
+    code execution is interrupted when the registered signals are encountered and - if provided -
+    a callback function is invoked.
+
+    After the callback, previous signal handlers are re-registered and the CM ends. If an exception
+    not related to the signal handling was encountered, it is re-raised after resetting the signal
+    handlers and (if set) invoking the callback function.
+
+    The callback function prototype is flexible, and the required arguments can be supplied by
+    the callback_args parameter. However, it should always accept the **kwargs arguments because
+    of the exc_type, exc_val and exc_tb arguments provided by the __exit__ function. This allows
+    the programmer to decide e.g. if certain parts of the callback code should be executed, based
+    on the raised exception - or the lack of an exception, that is.
+
+    A custom signal handler function can be supplied. In this case, the prototype should oblige
+    the rules of signal handling functions: func_name(signal_number, frame). If the custom signal
+    handling function uses a different exception then the default, it should be supplied to the CM
+    as well.
+
+    :ivar list signals: the list of signals that are being handled by the CM
+    :ivar function handler: the function used to handle the registered signals
+    :ivar exception handler_exc: the exception type related to the signal handler
+    :ivar function callback: the function that is always invoked during the CM exit
+    :ivar list callback_args: arguments for the callback function
+    :ivar list old_handlers: the list of previous signal handlers
+
+    """
+
+    __slots__ = ["signals", "handler", "handler_exc", "callback", "callback_args", "old_handlers"]
+
+    def __init__(self, *signals: int, **kwargs: Any) -> None:
+        """
+        :param signals: the identification of the handled signal, 'signal.SIG_' is recommended
+        :param kwargs: additional properties of the context manager
+        """
+        self.signals = signals
+        self.handler = kwargs.get("handler", common_kit.default_signal_handler)
+        self.handler_exc = kwargs.get("handler_exception", SignalReceivedException)
+        self.callback = kwargs.get("callback")
+        self.callback_args = kwargs.get("callback_args", [])
+        self.old_handlers: list[int | None | Callable[[int, Optional[types.FrameType]], Any]] = []
+
+    def __enter__(self) -> "HandledSignals":
+        """The CM entry sentinel, register the new signal handlers and store the previous ones.
+
+        :return object: the CM instance
+        """
+        for sig in self.signals:
+            self.old_handlers.append(signal.signal(sig, self.handler))
+        return self
+
+    def __exit__(self, exc_type: str, exc_val: Exception, exc_tb: traceback.StackSummary) -> bool:
+        """The CM exit sentinel, perform the callback and reset the signal handlers.
+
+        :param type exc_type: the type of the exception
+        :param exception exc_val: the value of the exception
+        :param traceback exc_tb: the traceback of the exception
+        :return bool: True if the encountered exception should be ignored, False otherwise or if
+                      no exception was raised
+        """
+        # Ignore all the handled signals temporarily
+        for sig in self.signals:
+            signal.signal(sig, signal.SIG_IGN)
+        # Perform the callback
+        if self.callback:
+            self.callback(*self.callback_args, exc_type=exc_type, exc_val=exc_val, exc_tb=exc_tb)
+        # Reset the signal handlers
+        for sig, sig_handler in zip(self.signals, self.old_handlers):
+            signal.signal(sig, sig_handler)
+        # Re-raise exceptions not related to signal handling done by the CM (e.g., SignalReceivedE.)
+        return isinstance(exc_val, self.handler_exc)
